Enhance version and imports.
