+++ restful_client_lite/__version__.py
@@ -1,0 +1,2 @@
+VERSION = (0, 0, 2)
+__version__ = ".".join(map(str, VERSION))
