+++ restful_client_lite/contrib/__init__.py
@@ -1,0 +1,1 @@
+import restful_client_lite.contrib.aliyun  # noqa
